+++ xlabuilder/literal.go
@@ -7,6 +7,8 @@ import "C"
 import (
 	"github.com/gomlx/exceptions"
 	"github.com/gomlx/gopjrt/dtypes"
+	"github.com/gomlx/gopjrt/dtypes/bfloat16"
+	"github.com/pkg/errors"
 	"github.com/x448/float16"
 	"reflect"
 	"runtime"
@@ -23,37 +25,54 @@ type Literal struct {
 
 // NewLiteralFromShape creates a zero-initialized literal with the given shape.
 // It cannot be used to create Literal tuples.
-func NewLiteralFromShape(shape Shape) *Literal {
+func NewLiteralFromShape(shape Shape) (*Literal, error) {
 	if shape.TupleSize() > 0 {
-		exceptions.Panicf("NewLiteralFromShape cannot be used to create tuple literals, shape given was %s", shape)
+		return nil, errors.Errorf("NewLiteralFromShape cannot be used to create tuple literals, shape given was %s", shape)
+	}
+	if shape.DType == dtypes.InvalidDType {
+		return nil, errors.Errorf("cannot create literal of invalid dtype, shape=%s", shape)
+	}
+	if shape.Size() <= 0 {
+		return nil, errors.Errorf("cannot create literal of size <= 0, shape=%s", shape)
 	}
 	cShape := cShapeFromShape(shape) // Ownership is given to the new Literal structure.
 	cLiteral := C.MakeLiteralFromShape(cShape)
-	return newLiteral(cLiteral, shape)
+	l, err := newLiteral(cLiteral, shape)
+	if err != nil {
+		return nil, err
+	}
+	return l, nil
 }
 
 // NewArrayLiteral creates a Literal initialized from the array flat data (a slice) and the dimensions of the array.
 //
 // If dimensions is omitted, it is assumed to represent a 1D-array of the length given.
-func NewArrayLiteral[T dtypes.Supported](flat []T, dimensions ...int) *Literal {
+func NewArrayLiteral[T dtypes.Supported](flat []T, dimensions ...int) (*Literal, error) {
 	if len(dimensions) == 0 {
 		dimensions = []int{len(flat)}
 	}
 	shape := MakeShape(dtypes.FromGenericsType[T](), dimensions...)
 	if shape.Size() != len(flat) {
-		exceptions.Panicf("NewArrayLiteral got a slice of length %d, but the shape %s given has %d elements",
+		return nil, errors.Errorf("NewArrayLiteral got a slice of length %d, but the shape %s given has %d elements",
 			len(flat), shape, shape.Size())
 	}
-	l := NewLiteralFromShape(shape)
+	l, err := NewLiteralFromShape(shape)
+	if err != nil {
+		return nil, err
+	}
 	lData := unsafe.Slice((*T)(unsafe.Pointer(l.cLiteral.data)), int(l.cLiteral.size))
 	copy(lData, flat)
-	return l
+	return l, nil
 }
 
 // NewScalarLiteral creates a scalar Literal initialized with the given value.
 func NewScalarLiteral[T dtypes.Supported](value T) *Literal {
 	shape := MakeShape(dtypes.FromGenericsType[T]())
-	l := NewLiteralFromShape(shape)
+	l, err := NewLiteralFromShape(shape)
+	if err != nil {
+		// Notice this should never happen, since dtypes.Support should always lead to valid shapes.
+		panic(err)
+	}
 	*(*T)(unsafe.Pointer(l.cLiteral.data)) = value
 	return l
 }
@@ -61,54 +80,74 @@ func NewScalarLiteral[T dtypes.Supported](value T) *Literal {
 // NewScalarLiteralFromFloat64 creates a scalar Literal with the given dtype initialized from the given value as float64.
 // This can be used to create common constants for arbitrary dtypes.
 //
-// It panics if the value cannot be converted to the dtype.
-func NewScalarLiteralFromFloat64(value float64, dtype dtypes.DType) *Literal {
-	// Scalar value.
-	if dtype == dtypes.Bool {
-		return NewScalarLiteral(value != 0)
+// It returns an error if dtype cannot be converted.
+func NewScalarLiteralFromFloat64(value float64, dtype dtypes.DType) (*Literal, error) {
+	// Scalar values.
+	switch dtype {
+	case dtypes.InvalidDType:
+		return nil, errors.Errorf("cannot create scalar literal of InvalidDtype, value=%g", value)
+	case dtypes.Bool:
+		return NewScalarLiteral(value != 0), nil
+	case dtypes.Complex64:
+		return NewScalarLiteral(complex(float32(value), 0)), nil
+	case dtypes.Complex128:
+		return NewScalarLiteral(complex(value, 0)), nil
+	case dtypes.Float16:
+		return NewScalarLiteral(float16.Fromfloat32(float32(value))), nil
+	case dtypes.BFloat16:
+		return NewScalarLiteral(bfloat16.FromFloat32(float32(value))), nil
+	default:
+		var convertedValue reflect.Value
+		panicMsg := exceptions.Try(func() {
+			convertedValue = reflect.ValueOf(value).Convert(dtype.GoType())
+		})
+		if panicMsg != nil {
+			return nil, errors.Errorf("failed to convert %g to dtype %s: %v", value, dtype, panicMsg)
+		}
+		l, err := NewLiteralFromShape(MakeShape(dtype))
+		if err != nil {
+			return nil, err
+		}
+		lValueOf := reflect.NewAt(dtype.GoType(), unsafe.Pointer(l.cLiteral.data)).Elem()
+		lValueOf.Set(convertedValue)
+		return l, nil
 	}
-	if dtype == dtypes.Complex64 {
-		return NewScalarLiteral(complex(float32(value), 0))
-	}
-	if dtype == dtypes.Complex128 {
-		return NewScalarLiteral(complex(value, 0))
-	}
-	if dtype == dtypes.Float16 {
-		return NewScalarLiteral(float16.Fromfloat32(float32(value)))
-	}
-
-	convertedValue := reflect.ValueOf(value).Convert(dtype.GoType())
-	l := NewLiteralFromShape(MakeShape(dtype))
-	lValueOf := reflect.NewAt(dtype.GoType(), unsafe.Pointer(l.cLiteral.data)).Elem()
-	lValueOf.Set(convertedValue)
-	return l
 }
 
 // NewScalarLiteralFromAny creates a scalar Literal with the given dynamically typed value.
 // It uses reflection to inspect the type.
-func NewScalarLiteralFromAny(value any) *Literal {
+func NewScalarLiteralFromAny(value any) (*Literal, error) {
 	valueOf := reflect.ValueOf(value)
 	dtype := dtypes.FromGoType(valueOf.Type())
-	l := NewLiteralFromShape(MakeShape(dtype))
+	if dtype == dtypes.InvalidDType {
+		return nil, errors.Errorf("Go type %T has no equivalent dtype", value)
+	}
+	l, err := NewLiteralFromShape(MakeShape(dtype))
+	if err != nil {
+		return nil, err
+	}
 	lValueOf := reflect.NewAt(dtype.GoType(), unsafe.Pointer(l.cLiteral.data)).Elem()
 	lValueOf.Set(valueOf)
-	return l
+	return l, nil
 }
 
 // NewArrayLiteralFromAny creates a scalar Literal with the given dynamically typed flat values and its underlying dimensions.
 // It uses reflection to inspect the type.
-func NewArrayLiteralFromAny(flatAny any, dimensions ...int) *Literal {
+func NewArrayLiteralFromAny(flatAny any, dimensions ...int) (*Literal, error) {
 	flatV := reflect.ValueOf(flatAny)
 	if flatV.Kind() != reflect.Slice {
-		exceptions.Panicf("NewArrayLiteralFromAny expects a slice, got %T instead", flatAny)
+		return nil, errors.Errorf("NewArrayLiteralFromAny expects a slice, got %T instead", flatAny)
 	}
 	dtype := dtypes.FromGoType(flatV.Type().Elem())
 	if dtype == dtypes.InvalidDType {
-		exceptions.Panicf("NewArrayLiteralFromAny expects a slice of valid DTypes, got %T instead", flatAny)
+		return nil, errors.Errorf("NewArrayLiteralFromAny expects a slice of valid DTypes, got %T instead", flatAny)
 	}
-	shape := MakeShape(dtype, dimensions...)
+	shape, err := MakeShapeOrError(dtype, dimensions...)
+	if err != nil {
+		return nil, err
+	}
 	if shape.Size() != flatV.Len() {
-		exceptions.Panicf("NewArrayLiteralFromAny got a slice of length %d, but the shape %s given has %d elements",
+		return nil, errors.Errorf("NewArrayLiteralFromAny got a slice of length %d, but the shape %s given has %d elements",
 			flatV.Len(), shape, shape.Size())
 	}
 
@@ -119,22 +158,24 @@ func NewArrayLiteralFromAny(flatAny any, dimensions ...int) *Literal {
 	defer pinner.Unpin()
 	flatData := unsafe.Slice((*byte)(flatPtr), shape.Memory())
 
-	l := NewLiteralFromShape(shape)
+	l, err := NewLiteralFromShape(shape)
+	if err != nil {
+		return nil, err
+	}
 	lData := unsafe.Slice((*byte)(unsafe.Pointer(l.cLiteral.data)), shape.Memory())
-
 	copy(lData, flatData)
-	return l
+	return l, nil
 }
 
 // newLiteral creates the literal and registers the finalizer.
-func newLiteral(cLiteral *C.Literal, shape Shape) *Literal {
+func newLiteral(cLiteral *C.Literal, shape Shape) (*Literal, error) {
 	l := &Literal{cLiteral: cLiteral, shape: shape}
 	if int(cLiteral.size) != shape.Size() {
-		exceptions.Panicf("new literal being created has shape=%s (size %d), but internal size reported is %d",
+		return nil, errors.Errorf("new literal being created has shape=%s (size %d), but internal size reported is %d",
 			shape, shape.Size(), cLiteral.size)
 	}
 	runtime.SetFinalizer(l, func(l *Literal) { l.Destroy() })
-	return l
+	return l, nil
 }
 
 // Destroy the Literal, release resources, and the Literal is no longer valid.